whatever

+++ src/samath.c
@@ -1,11 +1,11 @@
 #include "samath.h"
 
-int floor_to_muiltiple(long x, long m)
+int floor_to_muiltiple(int x, int m)
 {
     return x-(x%m);
 }
 
-int ceiling_to_muiltiple(long x, long m)
+int ceiling_to_muiltiple(int x, int m)
 {
     return x+(x%m);
 }
